Only places simple X marks on a Gameboard obj
Figuring out logic. I need to change the whole way the event listeners
work on the displayController. The display controller should have a
renderGameBoard() function that renders the whole board, and is called
everytime an action is done by players. It places the 'X' or 'O's based
on this rendering. This way the display Controller doesnt have to
worry about any logic unlike the way its coded right now.

Right now each square div's event listener has to keep track of a
reference to itself, since it has to check the Game Board in order to
decide whether or not it should display a marker. And we cant have the
gameboard require a div reference as a parameter bc then it wouldnt be
a standalone/module anymore. So... simply change it so the
displayController calls a renderGameBoard after every action