Basic working game logic on console
Now will work on adding wins to players, and im gonna try putting
players on an array instead of just p1 and p2, to simulate a situation
where id have more players for a diff game. Then just iterate with
array.find() to match the wnning symbols to the player with that symbol
in the array

Also changing name and symbol should NOT be a popup!!!. It should be
opt-in by simply clicking on a players name in order to change it.
This way, playing is quick, simple & easy. Yet if you want different
names, it is still easy to change them. Also, need to add a css class
that will highlight the current player. Also add wincounters etc